feat(repeat): add confirm dialog when trying to schedule repeated task for another day

+++ src/app/features/tasks/dialog-add-task-reminder/dialog-add-task-reminder.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { TaskService } from '../task.service';
 import { ReminderCopy } from '../../reminder/reminder.model';
 import { ReminderService } from '../../reminder/reminder.service';
@@ -9,6 +9,8 @@ import { throttle } from 'helpful-decorators';
 import { Task, TaskReminderOption, TaskReminderOptionId } from '../task.model';
 import { millisecondsDiffToRemindOption } from '../util/remind-option-to-milliseconds';
 import { LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG } from '../../../core/persistence/ls-keys.const';
+import { isToday } from '../../../util/is-today.util';
+import { DialogConfirmComponent } from '../../../ui/dialog-confirm/dialog-confirm.component';
 
 @Component({
   selector: 'dialog-add-task-reminder',
@@ -63,6 +65,7 @@ export class DialogAddTaskReminderComponent {
   constructor(
     private _taskService: TaskService,
     private _reminderService: ReminderService,
+    private _matDialog: MatDialog,
     private _matDialogRef: MatDialogRef<DialogAddTaskReminderComponent>,
     @Inject(MAT_DIALOG_DATA) public data: AddTaskReminderInterface,
   ) {
@@ -112,17 +115,44 @@ export class DialogAddTaskReminderComponent {
       });
       this.close();
     } else {
-      this._taskService.scheduleTask(
-        this.task,
-        timestamp,
-        this.selectedReminderCfgId,
-        this.isMoveToBacklog,
-      );
-      localStorage.setItem(
-        LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
-        this.isMoveToBacklog + '',
-      );
-      this.close();
+      if (!!this.task.repeatCfgId && !isToday(timestamp)) {
+        this._matDialog
+          .open(DialogConfirmComponent, {
+            restoreFocus: true,
+            data: {
+              message: T.F.TASK.D_REMINDER_ADD.CONFIRM_REPEAT_TXT,
+              okTxt: T.F.TASK.D_REMINDER_ADD.CONFIRM_REPEAT_OK,
+            },
+          })
+          .afterClosed()
+          .subscribe((isConfirm: boolean) => {
+            if (isConfirm) {
+              this._taskService.scheduleTask(
+                this.task,
+                timestamp,
+                this.selectedReminderCfgId,
+                this.isMoveToBacklog,
+              );
+              localStorage.setItem(
+                LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
+                this.isMoveToBacklog + '',
+              );
+              this.close();
+            }
+          });
+      } else {
+        this._taskService.scheduleTask(
+          this.task,
+          timestamp,
+          this.selectedReminderCfgId,
+          this.isMoveToBacklog,
+        );
+        localStorage.setItem(
+          LS_LAST_IS_MOVE_SCHEDULED_TO_BACKLOG,
+          this.isMoveToBacklog + '',
+        );
+        this.close();
+      }
     }
   }
